Report platform info when failure occurs during app/graph compilation
